Inital Commit
This is nowhere near where I want this.

This project is for fun little things.  Right now there is *no* effort to integrate it with the other parts of the project, but in the end it likely will be.

This is the start of writing a little app that calculates the value of the strength of group cuddles, or group meetings with people.

Much of this needs a lot of work done to it.

controller/*: this is the code which ends up getting executed first.

This likely will have a number of controllers listed overall, but only contains:

controller/cuddle-stress.php: cuddle stress value calculator.

This is the part of this project we are currently working on.  It calculates the little nifty value we are trying to get to caluclate here.

includes/*: these are files which do not really fit anywhere else (at least right now).

includes/verify.php: These are our data verification functions.

index.php: This displays basic stuff.  Not quite how I want it, but right now will do.  Displays content at the top, and the list of views with links to them.  Need to update to display controllers, but as no controller is working, will fix later.

view/*: files related to the display.  Format is:

controller-name.displaytype.php

This probably needs to be formatted a little differently in the future when different displaytypes might be created.

I think I am going to play with css display types before too long.

view/header.php: HTML header file.
view/footer.php: HTML footer file.
view/cuddle-stress.html.php: HTML view for Cuddle Stress controller.

I'm noticing my references are not *quite* right.

This is currently quite a lot of coding, but right now want to look at some of my convensions to make things clearer where I am heading with stuff.