Compteted Verify Routine
This calls a number of other routines, but this looks like it should mostly work for now.

There is an issue that we haven't set the valid values for the select.  Nor do we seem to have any way of dealing with the select not being selected at all, and if that is even allowed.