Added select verify function and some comments about return values
Select verify function has been added but is empty.

Initial comments about return values.  Need to update that.  The value returned in controller/cuddle-stress.php funtion verifyCuddleStressInput provides a blueprint for the return values I would like to have, though in these funtions there is no specific "verified" values which to return, but a generic of the type.